+++ src/navigation/common.py
@@ -12,14 +12,12 @@ from visualization_msgs.msg import Marker
 
 # TODO: rename?
 class Context:
-    is_shutdown: bool
     vel_cmd_publisher: rospy.Publisher
     vis_publisher: rospy.Publisher
     tf_buffer: tf2_ros.Buffer
     tf_listener: tf2_ros.TransformListener
 
     def __init__(self):
-        self.is_shutdown = False
         self.vel_cmd_publisher = rospy.Publisher('cmd_vel', Twist, queue_size=1)
         self.vis_publisher = rospy.Publisher('nav_vis', Marker)
         self.tf_buffer = tf2_ros.Buffer()
