+++ src/util/np_utils.py
@@ -3,6 +3,4 @@ import numpy as np
 
 def normalized(v):
     norm = np.linalg.norm(v)
-    if norm == 0:
-        norm = np.finfo(v.dtype).eps
     return v / norm
